Add a safe way to add render handlers for offscreen objects.
We rely on objects being able to run their animationy stuff even
when off screen so that they run to completion and remove themselves
from the screen. The Mesh.registerBeforeRender() ad Mesh.registerAfterRender()
do not run for Meshes that are occluded by the camera, so we have to
register these handlers with the scene directly instead. However, that
means keeping track of these handlers and making sure they get cleaned
up when the Mesh is removed from the Scene. The mesh-util module provides
functions to help with that.