Make the menu sort of work but ugly.
This adds new buttons that are emitted by joysticks to make
menu use with controllers possible.

The menu stuff is ridik and probably goes against react somehow
but it was the only way I could get it to work in a way that
made a slight amount of sense to me for now.